Rename home shell to app shell
