+++ presentation/app-shell/src/main/kotlin/com/sottti/roller/coasters/presentation/app/shell/model/AppShellActions.kt
@@ -1,0 +1,15 @@
+package com.sottti.roller.coasters.presentation.app.shell.model
+
+import androidx.compose.runtime.Immutable
+import com.sottti.roller.coasters.presentation.navigation.NavigationDestination
+
+@Immutable
+internal sealed interface AppShellActions {
+    @Immutable
+    data class DestinationSelected(
+        val destination: NavigationDestination,
+    ) : AppShellActions
+
+    @Immutable
+    data object NoOp : AppShellActions
+}
