+++ presentation/app-shell/src/main/res/values-gl/strings.xml
