+++ presentation/app-shell/src/test/kotlin/com/sottti/roller/coasters/presentation/app/shell/ui/AppShellUiBottomBarUiSnapshotTest.kt
@@ -1,0 +1,32 @@
+package com.sottti.roller.coasters.presentation.app.shell.ui
+
+import com.android.ide.common.rendering.api.SessionParams
+import com.android.resources.NightMode
+import com.sottti.roller.coasters.presentation.paparazzi.nightModeParameters
+import com.sottti.roller.coasters.presentation.paparazzi.paparazzi
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.junit.runners.Parameterized
+
+@RunWith(Parameterized::class)
+internal class AppShellUiBottomBarUiSnapshotTest(
+    nightMode: NightMode,
+) {
+
+    @get:Rule
+    val paparazzi = paparazzi(nightMode, SessionParams.RenderingMode.SHRINK)
+
+    @Test
+    fun snapshotTest() {
+        paparazzi.snapshot {
+            BottomBarPreview()
+        }
+    }
+
+    companion object Companion {
+        @JvmStatic
+        @Parameterized.Parameters
+        fun data(): Collection<Array<Any>> = nightModeParameters()
+    }
+}
