--- presentation/home/build.gradle.kts
@@ -1,37 +1,0 @@
-import com.sottti.roller.coasters.buildsrc.module
-
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.hilt)
-    alias(libs.plugins.kotlin.compose)
-    alias(libs.plugins.kotlin.ksp)
-    alias(libs.plugins.paparazzi)
-}
-
-android {
-    namespace = "com.sottti.roller.coasters.presentation.home"
-}
-
-dependencies {
-    implementation(libs.appcompat)
-    implementation(libs.compose.material)
-    implementation(libs.compose.navigation2)
-    implementation(libs.hilt)
-    implementation(libs.hilt.navigation.compose)
-    implementation(platform(libs.compose.bom))
-    implementation(project(module.domain.settings))
-    implementation(project(module.presentation.aboutMe))
-    implementation(project(module.presentation.designSystem.icons))
-    implementation(project(module.presentation.designSystem.text))
-    implementation(project(module.presentation.designSystem.themes))
-    implementation(project(module.presentation.explore))
-    implementation(project(module.presentation.favourites))
-    implementation(project(module.presentation.navigation))
-    implementation(project(module.presentation.previews))
-    implementation(project(module.presentation.rollerCoasterDetails))
-    implementation(project(module.presentation.search))
-    implementation(project(module.presentation.settings))
-    implementation(project(module.presentation.utils))
-    ksp(libs.hilt.compiler)
-    testImplementation(project(module.presentation.paparazzi))
-}
