--- presentation/home/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <activity
-            android:name=".ui.HomeActivity"
-            android:configChanges="uiMode"
-            android:exported="false"
-            android:icon="@mipmap/ic_launcher"
-            android:theme="@style/Theme.RollerCoasters" />
-    </application>
-
-</manifest>
