--- presentation/home/src/main/kotlin/com/sottti/roller/coasters/presentation/home/data/HomeMapper.kt
@@ -1,12 +1,0 @@
-package com.sottti.roller.coasters.presentation.home.data
-
-import com.sottti.roller.coasters.domain.settings.model.colorContrast.ResolvedColorContrast
-import com.sottti.roller.coasters.presentation.design.system.colors.color.ColorContrast
-
-internal fun ResolvedColorContrast.toColorContrast() =
-    when (this) {
-        ResolvedColorContrast.HighContrast -> ColorContrast.High
-        ResolvedColorContrast.MediumContrast -> ColorContrast.Medium
-        ResolvedColorContrast.StandardContrast -> ColorContrast.Standard
-        ResolvedColorContrast.LowContrast -> ColorContrast.Low
-    }
