--- presentation/home/src/main/kotlin/com/sottti/roller/coasters/presentation/home/model/HomeActions.kt
@@ -1,15 +1,0 @@
-package com.sottti.roller.coasters.presentation.home.model
-
-import androidx.compose.runtime.Immutable
-import com.sottti.roller.coasters.presentation.navigation.NavigationDestination
-
-@Immutable
-internal sealed interface HomeActions {
-    @Immutable
-    data class DestinationSelected(
-        val destination: NavigationDestination,
-    ) : HomeActions
-
-    @Immutable
-    data object NoOp : HomeActions
-}
