--- presentation/home/src/main/kotlin/com/sottti/roller/coasters/presentation/home/ui/HomeActivity.kt
@@ -1,41 +1,0 @@
-package com.sottti.roller.coasters.presentation.home.ui
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.sottti.roller.coasters.domain.settings.model.theme.ResolvedTheme
-import com.sottti.roller.coasters.presentation.design.system.themes.RollerCoastersTheme
-import com.sottti.roller.coasters.presentation.home.data.HomeViewModel
-import com.sottti.roller.coasters.presentation.home.data.toColorContrast
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-internal open class HomeActivity : AppCompatActivity() {
-
-    private val viewModel: HomeViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        enableEdgeToEdge()
-        super.onCreate(savedInstanceState)
-        setContent {
-            viewModel.state.collectAsStateWithLifecycle().value.let { state ->
-                RollerCoastersTheme(
-                    colorContrast = state.colorContrast.toColorContrast(),
-                    useDynamicColor = state.dynamicColor.enabled,
-                    useDarkTheme = state.theme == ResolvedTheme.DarkResolvedTheme,
-                ) {
-                    HomeUi(state = state, onAction = viewModel.onAction)
-                }
-            }
-        }
-    }
-}
-
-public fun startHomeActivity(context: Context) {
-    context.startActivity(Intent(context, HomeActivity::class.java))
-}
