+++ presentation/navigation/src/main/kotlin/com/sottti/roller/coasters/presentation/navigation/NavigationDestinationMappers.kt
@@ -1,9 +1,9 @@
 package com.sottti.roller.coasters.presentation.navigation
 
 import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.AboutMe
+import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.AppShell
 import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.Explore
 import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.Favourites
-import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.Home
 import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.Search
 import com.sottti.roller.coasters.presentation.navigation.NavigationDestination.Settings
 
@@ -12,7 +12,7 @@ public fun String?.toNavigationDestination(): NavigationDestination =
         AboutMe::class.qualifiedName -> AboutMe
         Explore::class.qualifiedName -> Explore
         Favourites::class.qualifiedName -> Favourites
-        Home::class.qualifiedName -> Home
+        AppShell::class.qualifiedName -> AppShell
         Search::class.qualifiedName -> Search
         Settings::class.qualifiedName -> Settings
         else -> Explore
